Fixed my mistakes

+++ src/main/java/com/madikhan/estore/action/impl/ajax/DeleteUserAction.java
@@ -1,21 +1,25 @@
 package com.madikhan.estore.action.impl.ajax;
 
-import static com.madikhan.estore.constants.NamesConstants.*;
 
 import com.madikhan.estore.action.Action;
 import com.madikhan.estore.exception.AccessDeniedException;
+import com.madikhan.estore.exception.BusinessException;
 import com.madikhan.estore.model.User;
 import com.madikhan.estore.service.OrderService;
 import com.madikhan.estore.service.UserService;
 import com.madikhan.estore.service.impl.OrderServiceImpl;
 import com.madikhan.estore.service.impl.UserServiceImpl;
 import com.madikhan.estore.util.SessionUtil;
+import com.madikhan.estore.validator.AdminValidator;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.sql.SQLException;
+
+import static com.madikhan.estore.constants.NamesConstants.*;
+
 
 public class DeleteUserAction implements Action {
 
@@ -26,13 +30,13 @@ public class DeleteUserAction implements Action {
     public void execute(HttpServletRequest request, HttpServletResponse response) throws IOException,
             SQLException, ServletException {
         User user = SessionUtil.getCurrentUser(request);
-        if (user.getIsAdmin()) {
+        if (AdminValidator.isAdminRole(request)) {
             Long userID = Long.valueOf(request.getParameter(USER_ID_ATTRIBUTE));
             Long countOrder = orderService.countUserOrders(userID);
-            if (countOrder == null || countOrder == ZERO) {
+            if ( (countOrder == null || countOrder == ZERO) && !user.getId().equals(userID)) {
                 userService.delete(userID);
             } else {
-                throw new Error();
+                throw new BusinessException(USER_HAS_ORDERS_MESSAGE);
             }
         } else {
             throw new AccessDeniedException(YOU_ARE_NOT_ADMIN_MESSAGE);
